:recycle:  moving P, useStableCallback, AnchorLink common components
Moving the code to separate files. Functionally should be no differences.
