+++ packages/desktop-client/src/components/common.tsx
@@ -1,21 +1,14 @@
-import React, {
-  useRef,
-  useLayoutEffect,
-  useCallback,
-  type ComponentProps,
-  type ReactNode,
-  forwardRef,
-} from 'react';
-import { NavLink, useMatch, useNavigate } from 'react-router-dom';
+import React, { type ComponentProps, type ReactNode, forwardRef } from 'react';
+import { useMatch, useNavigate } from 'react-router-dom';
 
-import { type CSSProperties, css } from 'glamor';
+import { type CSSProperties } from 'glamor';
 
-import { styles, colors } from '../style';
-import type { HTMLPropsWithStyle } from '../types/utils';
+import { colors } from '../style';
 
 import Button from './common/Button';
 
 export { default as AlignedText } from './common/AlignedText';
+export { default as AnchorLink } from './common/AnchorLink';
 export { default as Block } from './common/Block';
 export { default as Button, ButtonWithLoading } from './common/Button';
 export { default as Card } from './common/Card';
@@ -37,45 +30,6 @@ export { default as Text } from './common/Text';
 export { default as TextOneLine } from './common/TextOneLine';
 export { default as View } from './common/View';
 export { default as LinkButton } from './common/LinkButton';
-
-type UseStableCallbackArg = (...args: unknown[]) => unknown;
-
-export const useStableCallback = (callback: UseStableCallbackArg) => {
-  const callbackRef = useRef<UseStableCallbackArg>();
-  const memoCallback = useCallback(
-    (...args) => callbackRef.current && callbackRef.current(...args),
-    [],
-  );
-  useLayoutEffect(() => {
-    callbackRef.current = callback;
-  });
-  return memoCallback;
-};
-
-type AnchorLinkProps = {
-  to: string;
-  style?: CSSProperties;
-  activeStyle?: CSSProperties;
-  children?: ReactNode;
-};
-
-export function AnchorLink({
-  to,
-  style,
-  activeStyle,
-  children,
-}: AnchorLinkProps) {
-  let match = useMatch({ path: to });
-
-  return (
-    <NavLink
-      to={to}
-      {...css([styles.smallText, style, match ? activeStyle : null])}
-    >
-      {children}
-    </NavLink>
-  );
-}
 
 let externalLinkColors = {
   purple: colors.p4,
@@ -128,20 +82,6 @@ export function ButtonLink({
         navigate(to);
       }}
     />
-  );
-}
-
-type PProps = HTMLPropsWithStyle<HTMLDivElement> & {
-  isLast?: boolean;
-};
-export function P({ style, isLast, children, ...props }: PProps) {
-  return (
-    <div
-      {...props}
-      {...css(!isLast && { marginBottom: 15 }, style, { lineHeight: '1.5em' })}
-    >
-      {children}
-    </div>
   );
 }
