+++ packages/desktop-client/src/components/manager/subscribe/Bootstrap.tsx
@@ -6,7 +6,8 @@ import { loggedIn } from 'loot-core/src/client/actions/user';
 import { send } from 'loot-core/src/platform/client/fetch';
 
 import { colors } from '../../../style';
-import { View, Text, Button, P, ExternalLink } from '../../common';
+import { View, Text, Button, ExternalLink } from '../../common';
+import Paragraph from '../../common/Paragraph';
 
 import { useBootstrapped, Title } from './common';
 import { ConfirmPasswordForm } from './ConfirmPasswordForm';
@@ -52,19 +53,19 @@ export default function Bootstrap() {
   return (
     <View style={{ maxWidth: 450, marginTop: -30 }}>
       <Title text="Welcome to Actual!" />
-      <P style={{ fontSize: 16, color: colors.n2 }}>
+      <Paragraph style={{ fontSize: 16, color: colors.n2 }}>
         Actual is a super fast privacy-focused app for managing your finances.
         To secure your data, you’ll need to set a password for your server.
-      </P>
+      </Paragraph>
 
-      <P isLast style={{ fontSize: 16, color: colors.n2 }}>
+      <Paragraph isLast style={{ fontSize: 16, color: colors.n2 }}>
         Consider opening{' '}
         <ExternalLink to="https://actualbudget.org/docs/tour/">
           our tour
         </ExternalLink>{' '}
         in a new tab for some guidance on what to do when you’ve set your
         password.
-      </P>
+      </Paragraph>
 
       {error && (
         <Text
