+++ packages/desktop-client/src/components/modals/MergeUnusedPayees.js
@@ -6,7 +6,8 @@ import { send } from 'loot-core/src/platform/client/fetch';
 
 import { colors } from '../../style';
 import { Information } from '../alerts';
-import { View, Text, Modal, ModalButtons, Button, P } from '../common';
+import { View, Text, Modal, ModalButtons, Button } from '../common';
+import Paragraph from '../common/Paragraph';
 
 let highlightStyle = { color: colors.p5 };
 
@@ -89,7 +90,7 @@ export default function MergeUnusedPayees({
       {() => (
         <View style={{ padding: 20, maxWidth: 500 }}>
           <View>
-            <P style={{ marginBottom: 10, fontWeight: 500 }}>
+            <Paragraph style={{ marginBottom: 10, fontWeight: 500 }}>
               {payees.length === 1 ? (
                 <>
                   The payee <Text style={highlightStyle}>{payees[0].name}</Text>{' '}
@@ -118,7 +119,7 @@ export default function MergeUnusedPayees({
                   </ul>
                 </>
               )}
-            </P>
+            </Paragraph>
 
             <Information>
               Merging will remove the payee and transfer any existing rules to
