+++ packages/desktop-client/src/components/modals/PlaidExternalMsg.js
@@ -3,7 +3,8 @@ import React, { useState, useRef } from 'react';
 import AnimatedLoading from '../../icons/AnimatedLoading';
 import { colors } from '../../style';
 import { Error } from '../alerts';
-import { View, Text, Modal, Button, P, ModalButtons } from '../common';
+import { View, Text, Modal, Button, ModalButtons } from '../common';
+import Paragraph from '../common/Paragraph';
 
 function renderError(error) {
   return (
@@ -62,11 +63,11 @@ export default function PlaidExternalMsg({
     >
       {() => (
         <View>
-          <P style={{ fontSize: 15 }}>
+          <Paragraph style={{ fontSize: 15 }}>
             To link your bank account, you will be moved to your browser for
             enhanced security. Click below and Actual will automatically resume
             when you have given your bank’s credentials.
-          </P>
+          </Paragraph>
           {error && renderError(error)}
 
           {waiting ? (
