+++ packages/desktop-client/src/components/schedules/DiscoverSchedules.js
@@ -13,7 +13,8 @@ import useSelected, {
 import useSendPlatformRequest from '../../hooks/useSendPlatformRequest';
 import { colors } from '../../style';
 import { getParent } from '../../util/router-tools';
-import { View, Stack, ButtonWithLoading, P } from '../common';
+import { View, Stack, ButtonWithLoading } from '../common';
+import Paragraph from '../common/Paragraph';
 import { Page, usePageType } from '../Page';
 import { Table, TableHeader, Row, Field, SelectCell } from '../table';
 import DisplayId from '../util/DisplayId';
@@ -157,15 +158,15 @@ export default function DiscoverSchedules() {
 
   return (
     <Page title="Found schedules" modalSize={{ width: 850, height: 650 }}>
-      <P>
+      <Paragraph>
         We found some possible schedules in your current transactions. Select
         the ones you want to create.
-      </P>
-      <P>
+      </Paragraph>
+      <Paragraph>
         If you expected a schedule here and don’t see it, it might be because
         the payees of the transactions don’t match. Make sure you rename payees
         on all transactions for a schedule to be the same payee.
-      </P>
+      </Paragraph>
 
       <SelectedProvider instance={selectedInst}>
         <DiscoverSchedulesTable loading={isLoading} schedules={schedules} />
