+++ packages/desktop-client/src/components/schedules/PostsOfflineNotification.js
@@ -4,7 +4,8 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import { send } from 'loot-core/src/platform/client/fetch';
 
 import { colors } from '../../style';
-import { Text, P, Button, Stack } from '../common';
+import { Text, Button, Stack } from '../common';
+import Paragraph from '../common/Paragraph';
 import { Page } from '../Page';
 import DisplayId from '../util/DisplayId';
 
@@ -26,7 +27,7 @@ export default function PostsOfflineNotification() {
 
   return (
     <Page title="Post transactions?" modalSize="small">
-      <P>
+      <Paragraph>
         {payees.length > 0 ? (
           <Text>
             The {plural ? 'payees ' : 'payee '}
@@ -53,17 +54,17 @@ export default function PostsOfflineNotification() {
           syncing failed. In order to avoid duplicate transactions, we let you
           choose whether or not to create transactions for these schedules.
         </Text>
-      </P>
-      <P>
+      </Paragraph>
+      <Paragraph>
         Be aware that other devices may have already created these transactions.
         If you have multiple devices, make sure you only do this on one device
         or you will have duplicate transactions.
-      </P>
-      <P>
+      </Paragraph>
+      <Paragraph>
         You can always manually post a transaction later for a due schedule by
         selecting the schedule and clicking “Post transaction” in the action
         menu.
-      </P>
+      </Paragraph>
       <Stack
         direction="row"
         justify="flex-end"
